+++ src/refactorings/move-statement-down/move-statement-down.test.ts
@@ -431,6 +431,27 @@ function sayByeBye() {
     [cursor]{howAreYou}
   </>;
 }`
+      },
+      {
+        description: "JSX attributes",
+        code: `const Component = (...inputProps) => (
+  <input
+    [cursor]tw="w-full py-1 text-lg border border-none rounded"
+    css={{
+      color: darkred,
+      background: white
+    }}
+  />
+)`,
+        expected: `const Component = (...inputProps) => (
+  <input
+    css={{
+      color: darkred,
+      background: white
+    }}
+    [cursor]tw="w-full py-1 text-lg border border-none rounded"
+  />
+)`
       }
     ],
     async ({ code, expected }) => {