+++ src/refactorings/move-statement-up/move-statement-up.ts
@@ -47,6 +47,7 @@ function updateCode(
     ArrayExpression: visitPath,
     Literal: visitPath,
     JSXElement: visitPath,
+    JSXAttribute: visitPath,
     JSXExpressionContainer: visitPath
   });
 
@@ -151,6 +152,7 @@ function hasChildWhichMatchesSelection(
     ArrayExpression: visitPath,
     Literal: visitPath,
     JSXElement: visitPath,
+    JSXAttribute: visitPath,
     JSXExpressionContainer: visitPath
   });